Addressed some coverity issues

+++ src/hdf/fractalhead.c
@@ -328,7 +328,7 @@ static int directblockRead(struct READER *reader, struct DATAOBJECT *dataobject,
 
       default:
         mylog("FHDB type 1 unsupported values %08" PRIX64 " %" PRIX64 "\n",
-              unknown, ftell(reader->fhd) - 4);
+              unknown, (uint64_t)ftell(reader->fhd) - 4);
         return MYSOFA_UNSUPPORTED_FORMAT;
       }
